must fix random to decimal and test method bodies create

+++ src/gamblegamble/IRandomValueGenerator.java
@@ -4,8 +4,9 @@ public interface IRandomValueGenerator {
 
 	
 
-	public int getRandomNum();
-	public int getRandomNumInRange(int min, int max);
+	public double getRandomNum();
+	public double getRandomNumInRange(int min, int max);
 	public boolean getHitProbability(double p);
+	
 
 }
